git reset --soft HEAD~1

+++ backend/src/main/java/org/mapleland/maplelanbackserver/service/MapService.java
@@ -12,9 +12,12 @@ import org.mapleland.maplelanbackserver.dto.update.PriceUpdateRequest;
 import org.mapleland.maplelanbackserver.dto.update.ServerColorRequest;
 import org.mapleland.maplelanbackserver.enumType.Region;
 import org.mapleland.maplelanbackserver.enumType.alert.AlertStatus;
-import org.mapleland.maplelanbackserver.exception.NotFoundMapException;
-import org.mapleland.maplelanbackserver.exception.NotFoundMapTicketException;
-import org.mapleland.maplelanbackserver.exception.NotFoundUserException;
+import org.mapleland.maplelanbackserver.exception.coflict.CoolDownConflictException;
+import org.mapleland.maplelanbackserver.exception.notfound.jari.NotFoundMapException;
+import org.mapleland.maplelanbackserver.exception.notfound.jari.NotFoundMapTicketException;
+import org.mapleland.maplelanbackserver.exception.notfound.jari.NotFoundUserException;
+import org.mapleland.maplelanbackserver.exception.unauthorization.UserMismatchException;
+import org.mapleland.maplelanbackserver.filter.AdminCheckFilter;
 import org.mapleland.maplelanbackserver.jwtUtil.JwtUtil;
 import org.mapleland.maplelanbackserver.repository.*;
 import org.mapleland.maplelanbackserver.resolve.RegionResolver;
@@ -27,6 +30,7 @@ import org.springframework.stereotype.Service;
 
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
+import java.time.Duration;
 import java.util.*;
 import java.time.LocalDateTime;
 
@@ -37,15 +41,15 @@ import java.time.LocalDateTime;
 public class MapService {
 
     ModelMapper modelMapper = new ModelMapper();
-    private final jariRepository registerRepository;
-    private final userRepository userRepository;
-    private final UserService userActiveCheckService;
+    private final JariRepository registerRepository;
+    private final UserRepository userRepository;
     private final MapleMapRepository mapleMapRepository;
-    private final jariRepository jariRepository;
+    private final JariRepository jariRepository;
     private final MonsterDropItemRepository monsterDropItemRepository;
     private final DiscordDmService dmService;
     private final MapInterRestRepository interestRepository;
     private final UtilMethod utilMethod;
+    private final WebSocketService webSocketService;
 
     String message;
     @Value("${frontend.redirect-url}")
@@ -55,14 +59,14 @@ public class MapService {
     public void mapRegisterServiceMethod(JariCreatedRequest dto, String token) {
 
         int userId = JwtUtil.getUserId(token);
+        String discordId = JwtUtil.getDiscordId(token);
 
 
         //사용자 검색 -> 사용자 값 꺼내옴
         User user = userRepository.findByUserId(userId)
-                .orElseThrow(() -> new IllegalArgumentException("유저를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundUserException("유저를 찾을 수 없습니다."));
 
-        //사용자 벤 체크 -> False 벤
-        userActiveCheckService.userActiveCheck(user);
+
 
         // 지역 정보 추출
         Region region = RegionResolver.getRegionEnumByMapName(dto.getMapName());
@@ -73,14 +77,18 @@ public class MapService {
 
         // 등록 티켓 확인
         if (!user.getMapTicket()) {
-            throw new NotFoundMapTicketException("등록 티켓이 없습니다.");
+            throw new NotFoundMapTicketException("등록 티켓이 없습니다. \n 등록한 자리를 거래 완료 처리를 해야 합니다");
         }
 
-        // 테스트용 코드 (배포 시 false 설정해야 함)
-        user.setMapTicket(true);
+        String role = AdminCheckFilter.adminCheckFilter(discordId);
+
+        if(role.equals("ROLE_ADMIN")) user.setMapTicket(true);
+
+
+        else user.setMapTicket(false);
 
         // Entity 변환 및 등록
-        jari entity = modelMapper.map(dto, jari.class);
+        Jari entity = modelMapper.map(dto, Jari.class);
         entity.setUserMapId(null);
         entity.setArea(region);
         entity.setIsCompleted(false);
@@ -89,6 +97,8 @@ public class MapService {
 
         interRestUser(dto,userId);
         registerRepository.save(entity);
+        webSocketService.sendLatestPosts(userId);
+
 
     }
 
@@ -163,15 +173,15 @@ public class MapService {
                 dmService.sendToUser(discordId, selfMessage);
             }
         }
-
     }
+
     public AlertStatus MapInterRestServiceMethod(AlertRequest dto, String token) {
 
         return utilMethod.updateAlertInterest(dto,token);
     }
 
     public MapResponse searchMapsListKeyword(String keyword){
-//        List<MapDto> mapDtos = searchMapsByKeyword(keyword);
+
 
 
         List<DropItemResponse> dropItemResponses = monsterInfo(keyword);
@@ -186,11 +196,7 @@ public class MapService {
 
     public List<JariResponse> searchMapsByKeyword(String keyword) {
         PageRequest pageRequest = PageRequest.of(0, 100); // 첫 페이지, 100개
-        List<jari> results = jariRepository.findTop100ByMapNameWithUser(keyword,pageRequest);
-        System.out.println("🔍 검색된 결과 수: " + results.size());
-
-
-
+        List<Jari> results = jariRepository.findTop100ByMapNameWithUser(keyword,pageRequest);
 
 
         return results.stream()
@@ -240,7 +246,7 @@ public class MapService {
 
 
     public List<JariResponse> findByRegionTag(String keyword){
-        List<jari> byArea = jariRepository.findByArea(Region.valueOf(keyword));
+        List<Jari> byArea = jariRepository.findByArea(Region.valueOf(keyword));
 
 
 
@@ -275,7 +281,7 @@ public class MapService {
         LocalDateTime endTime = now;
 
         // 2. 최근 6시간 거래 조회 (isCompleted = true)
-        List<jari> completed =
+        List<Jari> completed =
                 jariRepository.findCompletedByMapNameIgnoreSpaceAndDate(keyword, startTime, endTime);
 
         // 3. 시간 슬롯 생성 (6개)
@@ -294,7 +300,7 @@ public class MapService {
 
             List<Integer> prices = completed.stream()
                     .filter(e -> !e.getCreateTime().isBefore(slotStart) && e.getCreateTime().isBefore(slotEnd))
-                    .map(jari::getPrice)
+                    .map(Jari::getPrice)
                     .toList();
 
             if (prices.isEmpty()) {
@@ -327,13 +333,19 @@ public class MapService {
         return filtered.stream().mapToInt(i -> i).average().orElse(0);
     }
 
-    public void mapUpdateAll(JariUpdateRequest jariUpdateRequest){
-        log.info("mapUpdateDto.mapId() = {}", jariUpdateRequest.mapId());
-        jari byUserMapId = registerRepository.findByUserMapId(jariUpdateRequest.mapId());
+    public void mapUpdateAll(JariUpdateRequest jariUpdateRequest,String token){
+
+        //jwt 유저 토큰 아이디
+        int userId = JwtUtil.getUserId(token);
+
+        Jari byUserMapId = registerRepository.findByUserMapId(jariUpdateRequest.mapId());
         if (byUserMapId == null) {
-            log.error("❌ 해당 mapId로 MapRegistrationEntity 를 찾을 수 없습니다: {}", jariUpdateRequest.mapId());
-            throw new RuntimeException("존재하지 않는 mapId입니다: " + jariUpdateRequest.mapId());
+            throw new NotFoundMapException("존재하지 않는 mapId입니다: " + jariUpdateRequest.mapId());
         }
+        Jari jariId = jariRepository.findByUserMapId(jariUpdateRequest.mapId());
+
+
+        if(userId != jariId.getUserMapId()) throw new UserMismatchException("사용자가 다릅니다.");
 
         byUserMapId.setServerColor(jariUpdateRequest.serverColor());
         byUserMapId.setPrice(jariUpdateRequest.price());
@@ -346,31 +358,51 @@ public class MapService {
 
 
     public void mapUpdatePrice(PriceUpdateRequest priceDto) {
-        jari byUserId = registerRepository.findByUserMapId(priceDto.mapId());
+        Jari byUserId = registerRepository.findByUserMapId(priceDto.mapId());
         byUserId.setPrice(priceDto.price());
         registerRepository.save(byUserId);
 
     }
 
     public void mapUpdateServerColor(ServerColorRequest dto) {
-        jari byUserId = registerRepository.findByUserMapId(dto.mapId());
+        Jari byUserId = registerRepository.findByUserMapId(dto.mapId());
         byUserId.setServerColor(dto.color());
         registerRepository.save(byUserId);
     }
 
-    public void mapDelete(int mapId) {
-        jari byUserId = registerRepository.findByUserMapId(mapId);
-        registerRepository.delete(byUserId);
+    public void mapDelete(int mapId,String token) {
+
+
+        Jari byUserId = registerRepository.findByUserMapId(mapId);
+        int userId = JwtUtil.getUserId(token);
+
+        if(byUserId.getUser().getUserId() == userId || JwtUtil.getRole(token).equals("ROLE_ADMIN")) {
+            registerRepository.delete(byUserId);
+        }
+        else throw new UserMismatchException("등록된 게시글과 다른 사용자 입니다.");
+
+
+
     }
 
-    public void updateIsCompleted(JariIsCompletedRequest dto) {
-        jari byUserMapId = registerRepository.findByUserMapId(dto.mapId());
+    public void updateIsCompleted(JariIsCompletedRequest dto , String token) {
+        int userId = JwtUtil.getUserId(token);
+        Jari jari = registerRepository.findByUserMapId(dto.mapId());
+        if(jari == null) throw new NotFoundMapException("게시글을 찾을 수 없습니다.");
 
-        if(byUserMapId == null) throw new NotFoundMapException("게시글을 찾을 수 없습니다.");
 
-        byUserMapId.setIsCompleted(true);
+        if(jari.getUser().getUserId() != userId) throw new UserMismatchException("사용자가 다릅니다.");
 
-        registerRepository.save(byUserMapId);
+        User user = userRepository.findByUserId(userId).
+                orElseThrow(() -> new NotFoundUserException("사용자를 찾을 수 없습니다."));
+
+        jari.setIsCompleted(true);
+        user.setMapTicket(true);
+
+
+
+        registerRepository.save(jari);
+        userRepository.save(user);
 
     }
 
@@ -386,5 +418,31 @@ public class MapService {
                 .toList();
 
         return new MapNameListResponse(MapNameList);
+    }
+
+    public void bumpJari(UserInformationService userInformationService, Integer jariId) {
+
+        Jari jari = jariRepository.findById(jariId).orElseThrow(RuntimeException::new);
+
+
+        if (!jari.validateOwner(userInformationService.getUserId())) {
+            throw new UserMismatchException("사용자가 아닙니다."); // 401
+        }
+
+        LocalDateTime now = LocalDateTime.now();
+        LocalDateTime createTime = jari.getCreateTime();
+        long minutesElapsed = Duration.between(createTime, now).toMinutes();
+        long remainingMinutes = 60 - minutesElapsed;
+        long remainingSeconds = Duration.between(createTime, now).getSeconds() % 60;
+
+
+        if (Duration.between(createTime, now).toMinutes() < 60) {
+            String message = String.format
+                    ("끌올은 한 시간에 한 번만 가능합니다. \n(%d분 %d초 후 다시 시도해주세요)", remainingMinutes, 60 - remainingSeconds);
+            throw new CoolDownConflictException(message); // 404
+        }
+
+        jari.bump(now);
+
     }
 }
